Added date parser, updated tests, new dependencies
1. Added a date parser - Parser receives a file handler from each file and read the first line.
If parser was able to identify the datetime format from known datetime formats (created as singleton to limit regex compilation to one time only), will return the offset (in chars).
If not successful, will skip the file.
Note: One of the reasons Python 3.7 was chosen was because it guarantees dict.keys() order.

2. Updated tests with various datetime formats
3. Included an unknown datetime format to be ignored in all tests
4. Two new Python dependencies (pip install singleton-decorator pytest-html)